+++ tests/dp_array_test.py
@@ -1,5 +1,4 @@
 """Tests the methods in DPArray."""
-import functools
 import numpy as np
 import pytest
 
@@ -152,39 +151,32 @@ def test_dtype_assignment(dtype):
     assert dp.dtype == dp.arr.dtype
 
 
-@pytest.mark.parametrize("n, funcs", [(10, [lambda x: x + 2, lambda x: x]),
-                                      (5, [lambda x: 2 * x, lambda x: x])],
-                         ids=["a", "b"])
-def test_max(n, funcs):
-    dp = DPArray(n)
+def test_max_empty_arrays_error():
+    dp = DPArray(10)
+    dp[0] = 0
+    dp[1] = 1
+    with pytest.raises(ValueError):
+        dp[2] = dp.max([], [])
+
+
+def test_max_arrays_size_error():
+    dp = DPArray(10)
+    dp[0] = 0
+    dp[1] = 1
+    with pytest.raises(ValueError):
+        dp[2] = dp.max([0, 1], [0])
+
+
+def test_max_cmp():
+    dp = DPArray(10)
     dp[0] = 0
     dp[1] = 1
 
-    with pytest.raises(
-            ValueError,
-            match="Expecting reference to be Iterable of length " + \
-            "at least one."
-        ):
-        dp[2] = dp.max(refs=2)
-
-    with pytest.raises(
-            ValueError,
-            match="Expecting reference to be Iterable of length " + \
-            "at least one."
-        ):
-        dp[2] = dp.max(refs=[])
-
-    with pytest.raises(
-            ValueError,
-            match="Expected refs and preprocessing of same length or " + \
-            "single preprocessing callable."
-    ):
-        dp[2] = dp.max(refs=[0, 1], preprocessing=[lambda x: x])
-
-    for i in range(2, n):
-        dp[i] = dp.max(refs=[i - 2, i - 1], preprocessing=funcs)
-        assert dp.arr[i] == max(funcs[0](dp.arr[i - 2]),
-                                funcs[1](dp.arr[i - 1]))
+    for i in range(2, 10):
+        indices = [i - 2, i - 1]
+        elements = [dp[indices[0]] + 2, 2 * dp[indices[1]]]
+        dp[i] = dp.max(indices, elements)
+        assert dp.arr[i] == max(*elements)
 
 
 def add_const(x, const):
@@ -194,12 +186,13 @@ def add_const(x, const):
     return x + const
 
 
+@pytest.mark.parametrize("max_min", ["max", "min"], ids=["max", "min"])
 @pytest.mark.parametrize(
-    "r, ans", [(np.array([[4, 3, 1], [5, 2, 1], [1, 2, 1]]), 14),
-               (np.array([[3, 4, 0, 0, 5], [4, 1, 2, 4, 4], [5, 1, 5, 5, 4],
-                          [2, 1, 1, 1, 4], [0, 0, 4, 3, 5]]), 36)],
-    ids=["a", "b"])
-def text_max_2d(r, truth):
+    "r, max_val, min_val",
+    [(np.array([[4, 3, 1], [5, 2, 1], [1, 2, 1]]), 14, 10),
+     (np.array([[3, 4, 0, 0, 5], [4, 1, 2, 4, 4], [5, 1, 5, 5, 4],
+                [2, 1, 1, 1, 4], [0, 0, 4, 3, 5]]), 36, 20)])
+def test_max_min_2d(max_min, r, max_val, min_val):
     """
     Given a reward matrix r, start at index (0, 0) (top left).
     Find the strategy yielding the largest reward when
@@ -209,49 +202,34 @@ def text_max_2d(r, truth):
     h, w = r.shape[0] + 1, r.shape[1] + 1
     dp = DPArray((h, w))
 
-    # fill column and row with padding
-    for i in range(w):
-        dp[0, i] = 0
-    for i in range(h):
-        dp[i, 0] = 0
+    if max_min == "max":
+        dp_cmp = dp.max
+        truth = max_val
+    else:
+        dp_cmp = dp.min
+        truth = min_val
 
-    # Note that dp and r indicies are off by one
-    for i, j in [(i, j) for i in range(1, h) for j in range(1, w)]:
-        func = functools.partial(add_const, const=r[i - 1, j - 1])
-        dp[i, j] = dp.max(refs=[(i - 1, j), (i, j - 1)], preprocessing=func)
+    # Base cases.
+    base_case = 0 if max_min == "max" else 1000
+    for i in range(w):
+        dp[0, i] = base_case
+    for i in range(h):
+        dp[i, 0] = base_case
+    dp[1, 1] = r[0, 0]
+
+    # Note that dp and r indicies are off by one.
+    for i in range(1, h):
+        for j in range(1, w):
+            if i == 1 and j == 1:
+                continue
+            indices = [(i - 1, j), (i, j - 1)]
+            elements = [
+                dp[indices[0]] + r[i - 1, j - 1],
+                dp[indices[1]] + r[i - 1, j - 1],
+            ]
+            dp[i, j] = dp_cmp(indices, elements)
 
     assert dp.arr[h - 1, w - 1] == truth
-
-
-@pytest.mark.parametrize(
-    "c, truth", [(np.array([[4, 3, 1], [5, 2, 1], [1, 2, 1]]), 10),
-                 (np.array([[3, 4, 0, 0, 5], [4, 1, 2, 4, 4], [5, 1, 5, 5, 4],
-                            [2, 1, 1, 1, 4], [0, 0, 4, 3, 5]]), 20)],
-    ids=["a", "b"])
-def test_min_2d(c, truth):
-    """
-    Given a cost matrix c, start at index (0, 0) (top left).
-    Find the strategy yielding the lowest cost path from the top
-    left of the matrix to the bottom left. Note that the user is
-    constrained to moving only down and right
-    The lowest cost achievable is given by truth
-    """
-    h, w = c.shape[0] + 1, c.shape[1] + 1
-    dp = DPArray((h, w))
-
-    # fill column and row with padding (Use a large number to pad)
-    for i in range(w):
-        dp[0, i] = 1000 * i
-    for i in range(h):
-        dp[i, 0] = 1000 * i
-    dp[1, 1] = c[0, 0]
-
-    # Note that dp and c indicies are off by one
-    for i, j in [(i, j) for i in range(1, h) for j in range(1, w)]:
-        func = functools.partial(add_const, const=c[i - 1, j - 1])
-        dp[i, j] = dp.min(refs=[(i - 1, j), (i, j - 1)], preprocessing=func)
-
-    assert dp.arr[h - 1, w - 1] - 1000 == truth
 
 
 # Logger related tests #